requirements_gemini.txt
google-generativeai>=0.5.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
requests>=2.31.0
